requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pyodbc==5.0.1
Werkzeug==3.0.1
fastapi
uvicorn
python-dotenv
pydantic
